requirements.txt
flask==3.0.0
openai==1.3.7
python-dotenv==1.0.0
qrcode==7.3
Pillow==10.1.0
SpeechRecognition==3.8.1
python-dateutil==2.8.2
flask-sqlalchemy==3.1.1
httpx>=0.24.1
